+++ src/transform.h
@@ -20,8 +20,8 @@ void quant(encoder_control* encoder, int16_t* pSrc, int16_t* pDes, int32_t iWidt
            int32_t iHeight, uint32_t *uiAcSum, int8_t eTType, int8_t scanIdx );
 void dequant(encoder_control* encoder, int16_t* piQCoef, int16_t* piCoef, int32_t iWidth, int32_t iHeight,int8_t eTType);
 
-void transform2d(int16_t *block,int16_t *coeff, int8_t blockSize, int8_t uiMode);
-void itransform2d(int16_t *block,int16_t *coeff, int8_t blockSize, int8_t uiMode);
+void transform2d(int16_t *block,int16_t *coeff, int8_t blockSize, int32_t uiMode);
+void itransform2d(int16_t *block,int16_t *coeff, int8_t blockSize, int32_t uiMode);
 
 void scalinglist_init();
 void scalinglist_processEnc( int32_t *coeff, int32_t *quantcoeff, int32_t quantScales, uint32_t height,uint32_t width, uint32_t ratio, int32_t sizuNum, uint32_t dc, uint8_t flat);
